Enable SiteAdmins JWT claim
This change will allow Site admins access to manage users and organisations.

+++ api/repositories/repositories.go
@@ -121,40 +121,62 @@ func GetRepository(c *fiber.Ctx) error {
 //Create a new repository
 func NewRepository(c *fiber.Ctx) error {
 	c.Accepts("application/json")
+	var id uuid.UUID
 	orgname := strings.ToLower(c.Params("org"))
 
-	user := c.Locals("user").(*jwt.Token)
-	claims := user.Claims.(jwt.MapClaims)
+	user_token := c.Locals("user").(*jwt.Token)
+	claims := user_token.Claims.(jwt.MapClaims)
 	username := claims["username"].(string)
+	// userorg_id := claims["userorg_id"].(uuid.UUID)
+	is_siteadmin := claims["site_admin"].(bool)
 	db := database.DB
-
+	// orgid := getOrganizationIDByUserName(orgname)
+	id = helpers.GetUserIDByUserName(username)
 	var organization models.Organization
 	repository := new(models.Repository)
-	db.Where("org_name = ?", orgname).Find(&organization)
+	// check JSON input
 	if err := c.BodyParser(repository); err != nil {
 		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Review your repository input", "Data": err})
 	}
+
+	// Check for a valid token
+	if !helpers.ValidToken(user_token, id) {
+		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Invalid token id", "data": nil})
+	}
+	// find the org in the organization table
+	db.Where("org_name = ?", orgname).Find(&organization)
 	if username != orgname {
 		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Unauthorized", "Data": nil})
 	}
-	db.Model(&organization).Where("org_name = ?", orgname).Association("Repositories").Append(repository)
+	if username == orgname || is_siteadmin {
+		db.Model(&organization).Where("org_name = ?", orgname).Association("Repositories").Append(repository)
+	}
 	return c.JSON(repository)
 }
 
 //Update a repository
 func UpdateRepository(c *fiber.Ctx) error {
 	c.Accepts("application/json")
+	var id uuid.UUID
 	orgname := c.Params("org")
 	reponame := c.Params("reponame")
 
-	user := c.Locals("user").(*jwt.Token)
-	claims := user.Claims.(jwt.MapClaims)
+	user_token := c.Locals("user").(*jwt.Token)
+	claims := user_token.Claims.(jwt.MapClaims)
 	username := claims["username"].(string)
+	is_siteadmin := claims["site_admin"].(bool)
 	userorg_id := claims["userorg_id"].(uuid.UUID)
 
 	db := database.DB
 	orgid := getOrganizationIDByUserName(orgname)
 	var repository models.Repository
+	if err := c.BodyParser(repository); err != nil {
+		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Review your repository input", "Data": err})
+	}
+
+	if !helpers.ValidToken(user_token, id) {
+		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Invalid token id", "data": nil})
+	}
 	if userorg_id != orgid || username != orgname {
 		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Unauthorized", "Data": nil})
 	}
@@ -168,17 +190,18 @@ func UpdateRepository(c *fiber.Ctx) error {
 
 	updatedRepository := new(models.Repository)
 
-	if err := c.BodyParser(updatedRepository); err != nil {
-		return apperr.BadRequest("Invalid params")
-	}
+	// if err := c.BodyParser(updatedRepository); err != nil {
+	// 	return apperr.BadRequest("Invalid params")
+	// }
 
 	updatedRepository = &models.Repository{Name: updatedRepository.Name, OrganizationID: updatedRepository.OrganizationID, Version: updatedRepository.Version, Description: updatedRepository.Description, URL: updatedRepository.URL}
 	updatedRepository.OrganizationID = orgid
 	updatedRepository.Name = reponame
-	if err = db.Where(&models.Repository{OrganizationID: orgid}).Model(&repository).Where("name = ?", reponame).Updates(updatedRepository).Error; err != nil {
-		return apperr.Unexpected(err.Error())
+	if userorg_id == orgid || username == orgname || is_siteadmin {
+		if err = db.Where(&models.Repository{OrganizationID: orgid}).Model(&repository).Where("name = ?", reponame).Updates(updatedRepository).Error; err != nil {
+			return apperr.Unexpected(err.Error())
+		}
 	}
-
 	return c.SendStatus(204)
 }
 
@@ -188,6 +211,7 @@ func DeleteRepository(c *fiber.Ctx) error {
 	user := c.Locals("user").(*jwt.Token)
 	claims := user.Claims.(jwt.MapClaims)
 	username := claims["username"].(string)
+	is_siteadmin := claims["site_admin"].(bool)
 	userorg_id, _ := uuid.Parse(claims["userorg_id"].(string))
 
 	db := database.DB
@@ -204,7 +228,8 @@ func DeleteRepository(c *fiber.Ctx) error {
 	} else if err != nil {
 		return apperr.Unexpected(err.Error())
 	}
-
-	db.Delete(&repository)
+	if userorg_id == orgid || username == orgname || is_siteadmin {
+		db.Delete(&repository)
+	}
 	return c.SendStatus(204)
 }
